requirements.txt
--index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://pypi.org/simple

ctranslate2==4.4.0
fastapi==0.115.5
faster_whisper==1.0.3
jiwer==3.0.4
notebook==7.2.2
numpy==1.23.5
pandas==2.2.3
piper-tts==1.2.0; platform_system == "Linux"
playsound==1.2.2
pyarrow==17.0.0
PyAudio==0.2.14
pydub==0.25.1
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
torch==2.3.1
torchaudio==2.3.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
uvicorn==0.32.0
